% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/rpart.R
\name{active_predictors}
\alias{active_predictors}
\alias{active_predictors.rpart}
\title{Extract the active features from a tree}
\usage{
active_predictors(x, ...)

\method{active_predictors}{rpart}(x, ...)
}
\arguments{
\item{x}{A object}

\item{...}{Other arguments passed to methods}
}
\value{
A tibble with list column \code{active_predictors} containing a character
vector of predictors.
}
\description{
If a tree does not use a predictor in the training set in any of its splits
it is functionally independent of the prediction function. This generic
returns a data frame containing character vector of predictor names that
were used in at least one split.
}
\examples{
if (rlang::is_installed(c("rpart", "palmerpenguins"))) {
  data(penguins, package = "palmerpenguins")
  penguins <- na.omit(penguins)

  # Fit a tree
  tree <- rpart::rpart(species ~ ., data = penguins)
  tree

  # Extract active predictors
  active_predictors(tree)

  # Only primary splits are included - competing and surrogate splits
  # are excluded since they don't affect predictions
}

}
