% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extract_rules}
\alias{extract_rules}
\title{Extract an expression that defines a path to a terminal node}
\usage{
extract_rules(x, ...)
}
\arguments{
\item{x}{A object}

\item{...}{Other arguments passed to methods}
}
\value{
A data frame with column \code{equation} (an R expression) and \code{id} (a
character identifier).
}
\description{
A rule is a logical expression of predictor variables that reflects which
data are contained in or sent to a terminal node in a tree-based model. Rules
can take any form but, for most trees, they are simple statements such as
\code{x < 1.2}, \code{y == "red"}, or \code{z \%in\% c("blue", "green")}.
}
