% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorax-package.R
\docType{package}
\name{lorax-package}
\alias{lorax}
\alias{lorax-package}
\title{lorax: Speak for the Trees}
\description{
Extracts decision rules from tree- and rule-based models fitted in 'R'. Rules are expressed as logical predicates that identify paths to terminal nodes, making model behavior more transparent and interpretable. Provides conversion methods to 'partykit' party objects for a wide range of model types, including those from 'ranger', 'xgboost', 'lightgbm', 'randomForest', 'C50', 'rpart', and 'dbarts'. The 'partykit' infrastructure is described in Hothorn and Zeileis (2015) \url{https://jmlr.org/papers/v16/hothorn15a.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/topepo/lorax}
  \item Report bugs at \url{https://github.com/topepo/lorax/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
