% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorax-package.R
\docType{package}
\name{lorax-package}
\alias{lorax}
\alias{lorax-package}
\title{lorax: Speak for the Trees}
\description{
R functions and methods to extract elements of tree- and rule-based models.
}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
