% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObliqueForest.R, R/grf.R, R/ranger.R, R/rpart.R
\name{var_imp.ObliqueForest}
\alias{var_imp.ObliqueForest}
\alias{lorax_var_imp}
\alias{var_imp.grf}
\alias{var_imp.ranger}
\alias{var_imp.rpart}
\title{Tree Importance Scores}
\usage{
\method{var_imp}{ObliqueForest}(object, complete = TRUE, ...)

\method{var_imp}{grf}(object, complete = TRUE, ...)

\method{var_imp}{ranger}(object, complete = TRUE, ...)

\method{var_imp}{rpart}(object, complete = TRUE, ...)
}
\arguments{
\item{object}{A model object.}

\item{complete}{A logical to filling absent importance values with zeros.}

\item{...}{Arguments passed to importance functions (if any).}
}
\value{
A tibble with columns \code{term} and \code{estimate}.
}
\description{
Methods for computing variable importance scores via the model object using
a common interface.
}
