% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorax-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.party}
\alias{var_imp}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{var_imp}}}

  \item{partykit}{\code{\link[partykit:party-coercion]{as.party}}}
}}

